# Improve performance of Path.ChangeExtension