--- a/src/System.Private.CoreLib/shared/System/String.Manipulation.cs
+++ b/src/System.Private.CoreLib/shared/System/String.Manipulation.cs
@@ -341,6 +341,55 @@ namespace System
             return result;
         }
 
+        // TODO: Expose publicly: https://github.com/dotnet/corefx/issues/34330.
+        internal static string Concat(ReadOnlySpan<char> str0, ReadOnlySpan<char> str1)
+        {
+            string result = FastAllocateString(str0.Length + str1.Length);
+            Span<char> resultSpan = new Span<char>(ref result.GetRawStringData(), result.Length);
+
+            str0.CopyTo(resultSpan);
+            str1.CopyTo(resultSpan.Slice(str0.Length));
+
+            return result;
+        }
+
+        // TODO: Expose publicly: https://github.com/dotnet/corefx/issues/34330.
+        internal static string Concat(ReadOnlySpan<char> str0, ReadOnlySpan<char> str1, ReadOnlySpan<char> str2)
+        {
+            string result = FastAllocateString(str0.Length + str1.Length + str2.Length);
+            Span<char> resultSpan = new Span<char>(ref result.GetRawStringData(), result.Length);
+
+            str0.CopyTo(resultSpan);
+            resultSpan = resultSpan.Slice(str0.Length);
+
+            str1.CopyTo(resultSpan);
+            resultSpan = resultSpan.Slice(str1.Length);
+
+            str2.CopyTo(resultSpan);
+
+            return result;
+        }
+
+        // TODO: Expose publicly: https://github.com/dotnet/corefx/issues/34330.
+        internal static string Concat(ReadOnlySpan<char> str0, ReadOnlySpan<char> str1, ReadOnlySpan<char> str2, ReadOnlySpan<char> str3)
+        {
+            string result = FastAllocateString(str0.Length + str1.Length + str2.Length + str3.Length);
+            Span<char> resultSpan = new Span<char>(ref result.GetRawStringData(), result.Length);
+
+            str0.CopyTo(resultSpan);
+            resultSpan = resultSpan.Slice(str0.Length);
+
+            str1.CopyTo(resultSpan);
+            resultSpan = resultSpan.Slice(str1.Length);
+
+            str2.CopyTo(resultSpan);
+            resultSpan = resultSpan.Slice(str2.Length);
+
+            str3.CopyTo(resultSpan);
+
+            return result;
+        }
+
         public static string Concat(params string[] values)
         {
             if (values == null)