@stephentoub 2019-01-02
@dotnet-bot test Ubuntu x64 Checked Innerloop Build and Test please
@dotnet-bot test Windows_NT arm64 Cross Debug Innerloop Build please