@jkotas 2019-01-02
Is this important enough method to add extra more complex code to avoid the allocations?

The Range/Index fix was close to my threashold, but I think this one is below it.