@stephentoub 2019-01-02
> Is this important enough method to add extra more complex code to avoid the allocations?

It may not be.  I can close if you don't think it's worth it.  I'd searched msbuild to see its usage (https://github.com/Microsoft/msbuild/search?p=1&q=ChangeExtension&unscoped_q=ChangeExtension) and some of those seemed like they might be meaningful, but no smoking gun. (Also http://index/?leftProject=mscorlib&leftSymbol=vf50vlj9fsry&file=system%5Cio%5Cpath.cs for folks with access.)