@jkotas 2019-01-03
This change would look fine to me if it used `Concat` that takes Spans. Do we have API issue for that? 

I think `Concat` that takes Spans would be useful. I have run myself into situations where it would be useful to have. Add the `Concat` as internal for now to support this change, and add API proposal to add it as public in parallel?