@stephentoub 2019-01-03
> Add the Concat as internal for now to support this change, and add API proposal to add it as public in parallel?

I was thinking the same thing. Will do.