@stephentoub 2019-01-03
I added String.Concat overloads for spans, utilized them in a few places in corelib (including in Path.ChangeExtension), and opened https://github.com/dotnet/corefx/issues/34330 to make the methods public.